requirements.txt
# =============================================================================
# SENTINEL — Python Dependencies
# =============================================================================

# Data & API
requests>=2.31.0
httpx>=0.27.0
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
alembic>=1.13.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Scheduling & Pipeline
prefect>=2.14.0

# Progress & Logging
tqdm>=4.66.0
loguru>=0.7.2

# Analysis & Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
statsmodels>=0.14.0
scipy>=1.12.0
scikit-learn>=1.4.0

# ML Models
torch>=2.2.0
pytorch-forecasting>=1.0.0
pytorch-lightning>=2.2.0
xgboost>=2.0.0
prophet>=1.1.5

# NLP
transformers>=4.38.0
feedparser>=6.0.11
newspaper3k>=0.2.8

# Graph Analysis
networkx>=3.2.0

# Dashboard
streamlit>=1.31.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
